+++ geomesa-z3/src/main/scala/org/locationtech/geomesa/curve/XZ3SFC.scala
@@ -62,7 +62,7 @@ class XZ3SFC(val g: Short, val xBounds: (Double, Double), val yBounds: (Double, 
     val l1 = math.floor(math.log(maxDim) / XZSFC.LogPointFive).toInt
 
     // the length will either be (l1) or (l1 + 1)
-    val length = if (l1 >= g) { g } else {
+    val length = if (l1 >= g) { g.toInt } else {
       val w2 = math.pow(0.5, l1 + 1) // width of an element at resolution l2 (l1 + 1)
 
       // predicate for checking how many axis the polygon intersects
